requirements.txt
anthropic[bedrock,vertex]>=0.37.1
Pillow>=10.0.0
PyAutoGUI>=0.9.54